+++ pgjdbc/src/main/java/org/postgresql/jdbc/PgResultSet.java
@@ -1357,7 +1357,12 @@ public class PgResultSet implements ResultSet, org.postgresql.PGRefCursorResultS
       PgResultSet rs = (PgResultSet) selectStatement.executeQuery();
 
       if (rs.next()) {
-        rowBuffer = rs.thisRow;
+        // we know that the row is updatable as it was tested above.
+        if ( rs.thisRow == null ) {
+          rowBuffer = null;
+        } else {
+          rowBuffer = castNonNull(rs.thisRow).updateableCopy();
+        }
       }
 
       castNonNull(rows).set(currentRow, castNonNull(rowBuffer));
@@ -1618,7 +1623,7 @@ public class PgResultSet implements ResultSet, org.postgresql.PGRefCursorResultS
     @Nullable String[] s = quotelessTableName(castNonNull(tableName));
     String quotelessTableName = castNonNull(s[0]);
     @Nullable String quotelessSchemaName = s[1];
-    java.sql.ResultSet rs = connection.getMetaData().getPrimaryKeys("",
+    java.sql.ResultSet rs = ((PgDatabaseMetaData)connection.getMetaData()).getPrimaryUniqueKeys("",
         quotelessSchemaName, quotelessTableName);
 
     while (rs.next()) {