+++ pgjdbc/src/test/java/org/postgresql/test/jdbc2/UpdateableResultTest.java
@@ -56,6 +56,8 @@ public class UpdateableResultTest extends BaseTest4 {
     // put some dummy data into second
     st2.execute("insert into second values (1,'anyvalue' )");
     st2.close();
+    TestUtil.createTable(con, "uniqueconstraint", "u1 int unique, name1 text");
+    TestUtil.execute("insert into uniqueconstraint values (1, 'dave')", con);
 
   }
 
@@ -68,6 +70,7 @@ public class UpdateableResultTest extends BaseTest4 {
     TestUtil.dropTable(con, "stream");
     TestUtil.dropTable(con, "nopkmulticol");
     TestUtil.dropTable(con, "booltable");
+    TestUtil.dropTable(con, "uniqueconstraint");
     super.tearDown();
   }
 
@@ -445,6 +448,26 @@ public class UpdateableResultTest extends BaseTest4 {
   }
 
   @Test
+  public void test2193() throws Exception {
+    Statement st =
+        con.createStatement(ResultSet.TYPE_SCROLL_INSENSITIVE, ResultSet.CONCUR_UPDATABLE);
+    ResultSet rs = st.executeQuery("select * from updateable");
+    assertNotNull(rs);
+    rs.moveToInsertRow();
+    rs.updateInt(1, 1);
+    rs.updateString(2, "jake");
+    rs.updateString(3, "avalue");
+    rs.insertRow();
+    rs.first();
+
+    rs.updateString(2, "bob");
+    rs.updateRow();
+    rs.refreshRow();
+    rs.updateString(2, "jake");
+    rs.updateRow();
+  }
+
+  @Test
   public void testInsertRowIllegalMethods() throws Exception {
     Statement st =
         con.createStatement(ResultSet.TYPE_SCROLL_INSENSITIVE, ResultSet.CONCUR_UPDATABLE);
@@ -700,6 +723,18 @@ public class UpdateableResultTest extends BaseTest4 {
     assertTrue(rs.next());
     assertTrue(rs.first());
     rs.updateString("relname", "pg_class");
+    rs.close();
+    st.close();
+  }
+
+  @Test
+  public void testUniqueUpdatable() throws Exception {
+    Statement st = con.createStatement(ResultSet.TYPE_SCROLL_INSENSITIVE,
+        ResultSet.CONCUR_UPDATABLE);
+    ResultSet rs = st.executeQuery("SELECT * from uniqueconstraint");
+    assertTrue(rs.next());
+    assertTrue(rs.first());
+    rs.updateString("name1", "bob");
     rs.close();
     st.close();
   }